requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
supabase==2.0.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
openai==1.3.0
transformers==4.35.0
torch==2.1.0
pandas==2.1.3
numpy==1.26.2
Pillow>=10.1.0
qrcode==7.4.2
reportlab==4.0.7
openpyxl==3.1.2
selenium==4.15.2
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
django-environ==0.11.2
requests==2.32.5

# Added by project scaffolding: optional storage/backends
django-storages>=1.13.1
boto3>=2.0.0

